Merged PR 2146163: Added task_group and removed C++/WinRT dependency
The task_group class is a simple abstraction that supports massive parallelism. This version provides a very efficient Windows thread pool implementation, but the task_group public interface is very simple, so a non-Windows implementation can be written. This avoids the coroutine dependency, which was the bulk of what the xlang tools relied on from C++/WinRT.

The remaining C++/WinRT dependencies have been removed. Work still needs to be done to make it actually cross-compile but at least now it should be relatively simple.

I didn't bother to remove the C++/WinRT dependency from the compare project since that is very much a Windows-only compare tool that we can probably scrap eventually.