Merged PR 2154486: Start of modern C++ projection
I started this C++/WinRT replacement on Friday. The idea is to get to a point where it generates the same output as the current cppwinrt compiler and then retire the existing compiler.

I also used my own scratch vcxproj file so that I can use a pch and it immediately picked up a bunch of missing `inline` keywords.