Merged PR 2292084: More progress on C++ projection
The C++ projection is now mostly complete. All of the platform's namespace headers are now generated mostly on par with cppwinrt. It doesn't all compile but it's now very close to the final form and all of the "hard" problems have been solved. Things like recursive interfaces, factories, and using declarations are all now folded and handled correctly. I've done a lot of work to simplify the code and developed helpers like method_signature, get_factories, and get_interfaces that are used heavily to solve various difficult query problems.