Design note cleanup (#14)
I've started the cleanup process for design notes. This PR renames the design_notes folder to be snake_case and updates XDN00 - 02 with new project naming. Additionally, I expanded the design note process doc to better describe the role of design notes relative to project documentation.

LDN03-06 remain unchanged for now. My plan is to move them into a docs folder and remove the design note elements in favor of a more project document style as a seperate PR

* rename DesignNotes folder to design_notes
* Update design note template to new project name
* renamed and updated XDN00 - design note process
* Update XDN01 - language interop strategy
* renamed LDN02 -> XDN02