writerbase::count_placeholders miscounts "^^%" as a placeholder
The sequence "^^%" is an escaped escape character followed by
a placeholder. However, count_placeholders mistakenly applied
the second ^ character to the % and concluded that the % was
escaped.

The character after an escape should turn off escape mode,
even if that character is itself an escape.

Test:
    struct writer : xlang::text::writer_base<writer> {} w;
    w.write("^^%", 1);

Without this fix, the code fails the expected/actual assertion.