writer_base::write() should disallow format strings ending with escape
If a format string ends with an escape, the code tried to
access the character after the format string, which was a range
violation on the string view.

We declare format strings ending with escape as explicitly illegal
and assert as such before triggering UB, thereby making the problem
more diagnosable.

Test:

    struct writer : xlang::text::writer_base<writer> {} w;
    w.write("%^^", 1);

Previous code would trigger UB trying to read past the end
of the string view. The fix asserts so you know what you did wrong.