writer_base:write() mishandles double escape characters
The format string "^^" is an escaped escape and therefore
should print one escape character, but instead it prints two.

Even worse: Given the sequence "^^%", the second escape
character remains live and protects the % character,
which it should not.

Test:

    struct writer : xlang::text::writer_base<writer> {} w;
    w.write("%^^", 1);
    w.write("^^%", 1);

Previous code would print two ^ characters, and the
second format would trigger an assertion because the
% was erroneously protected.