cache.remove_cppwinrt_foundation_types didn't remove everything
Forgot to remove the types from the member.types map.

And if the `Windows.Foundation` namespace didn't exist,
remove_cppwinrt_foundation_types would accidentally create it.
You don't expect calling a "remove" function to create stuff!

Note that if we remove the last type from a namespace, we
do not remove the namespace.