Avoid runaway memory usage if blob has absurd element count
The number of elements cannot be larger than the blob size
because each element is at least 1 byte in size. We would
discover the problem eventually when we ran out of data,
but not until after we had allocated 4GB of memory, resulting
in a potential DoS against the build machine.