TypeDef.Extends() could be a TypeDef or a TypeRef
According to the spec, [TypeDef redirection is optional(https://github.com/Microsoft/xlang/blob/master/docs/XSPEC02%20-%20Metadata%20Representation.md#typedef-redirection):

> 3rd party metadata files may use TypeDef directly or may redirect all type references thru a typeRef similar to how system metadata files do.

Despite its name, TypeDefOrRef could be one of *four* things:

* TypeDef
* TypeRef
* TypeSpec
* Nothing

TypeSpec is used for generic types, but deriving from a generic type is not allowed, so you won't see that in an Extends().

Nothing is used only for System.Object, because that is the only type with no base class.

TypeDef and TypeRef are in play, and both are allowed.

There's now a circular dependency between TypeDef and extends_type, which I broke by changing the return values to be explicit rather than auto.

This addresses issue #249