Improve scalability of co_await apartment_context
When we fixed throughput of coroutine resumption in #546, we failed to
update the corresponding code in apartment_context.  Refactored the
logic into a common function and have everybody use it.

Also make the thread_pool test more reliable. Sometimes fails spuriously
because the envisioned race condition fails to materialize. Make the
race condition far more likely to occur.