Fix RoResolveNamespace not looking at exe directory on RS5 as fallback.  (#676)
Types were failing to resolve on RS5 because RoResolveNamespace did not look at the exe file path as fallback by default. Supplying the exe file path to windowsMetaData dir only looks for types with the Windows prefix. The fix was to also supply the exe file path to the package graph parameter.

Also fix xml parsing to look for 'name' on ActivatableClass tag instead of 'clsid'.