Embedded Manifest (#707)
Implemented Embedded Manifests for exe and dll

To enable dll embedded manifests, I call into CreateActCtxW and QueryActCtxW and iterate through the manifests assemblies.

I've also implemented an embedded manifest tests which works exactly like the UndockedRegFreeWinRTTest.exe but the exe manifest is embedded and it has a dependentAssembly on the EmbeddedTestComponent. It looks at the manifest that is embedded in EmbeddedTestComponent.dll and activates the classes from it.

There are also structural improvements to the project.